Added sub-categories and moved some settings to more appropriate categories
